# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Academia Rest API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate a user",
                "description": "Authenticates a user by providing username and password.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admins2222"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a list of users",
                "description": "Returns a list of users",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter criteria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search criteria",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Endpoint to create a new user",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "user_type": {
                                        "type": "string",
                                        "enum": [
                                            "student",
                                            "teacher",
                                            "admin",
                                            "driver"
                                        ]
                                    },
                                    "username": {
                                        "type": "string",
                                        "maxLength": 10
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AddressResource": {
                "title": "Address Data",
                "description": "Data for address-related data",
                "properties": {
                    "user_id": {
                        "description": "User ID associated with the address",
                        "type": "integer"
                    },
                    "address_type": {
                        "description": "Type of address (e.g., home, work)",
                        "type": "string"
                    },
                    "address_line_1": {
                        "description": "First line of address",
                        "type": "string"
                    },
                    "address_line_2": {
                        "description": "Second line of address",
                        "type": "string"
                    },
                    "display": {
                        "description": "Formatted display of the address",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Data",
                "description": "Schema for user-related data",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "username": {
                        "description": "User's username",
                        "type": "string"
                    },
                    "user_type": {
                        "description": "Type of user",
                        "type": "string"
                    },
                    "addresses": {
                        "description": "User's addresses",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddressResource"
                        },
                        "nullable": true
                    },
                    "details": {
                        "description": "User's details",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "cookie",
                "scheme": "bearer"
            }
        }
    }
}